padcli/command/testdata/requirements.txt
uvicorn
fastapi
jetpack-io>=0.5.1.dev202203101646946286
